@@ -15,62 +15,62 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.9",
-    "@angular/common": "~12.2.9",
-    "@angular/compiler": "~12.2.9",
-    "@angular/core": "~12.2.9",
-    "@angular/forms": "~12.2.9",
-    "@angular/platform-browser": "~12.2.9",
-    "@angular/platform-browser-dynamic": "~12.2.9",
-    "@angular/router": "~12.2.9",
-    "@fortawesome/angular-fontawesome": "^0.9.0",
+    "@angular/animations": "~13.0.1",
+    "@angular/common": "~13.0.1",
+    "@angular/compiler": "~13.0.1",
+    "@angular/core": "~13.0.1",
+    "@angular/forms": "~13.0.1",
+    "@angular/platform-browser": "~13.0.1",
+    "@angular/platform-browser-dynamic": "~13.0.1",
+    "@angular/router": "~13.0.1",
+    "@fortawesome/angular-fontawesome": "^0.10.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-brands-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@ngrx/component": "^12.4.0",
-    "@ngrx/effects": "^12.4.0",
-    "@ngrx/router-store": "^12.4.0",
-    "@ngrx/store": "^12.4.0",
-    "@ngrx/store-devtools": "^12.4.0",
-    "@taiga-ui/cdk": "^2.19.0",
-    "@taiga-ui/core": "^2.19.0",
-    "@taiga-ui/icons": "^2.19.0",
-    "@taiga-ui/kit": "^2.19.0",
-    "rxjs": "~6.6.0",
-    "tailwindcss": "^2.2.16",
+    "@ngrx/component": "^12.5.1",
+    "@ngrx/effects": "^12.5.1",
+    "@ngrx/router-store": "^12.5.1",
+    "@ngrx/store": "^12.5.1",
+    "@ngrx/store-devtools": "^12.5.1",
+    "@taiga-ui/cdk": "^2.22.0",
+    "@taiga-ui/core": "^2.22.0",
+    "@taiga-ui/icons": "^2.22.0",
+    "@taiga-ui/kit": "^2.22.0",
+    "rxjs": "^7.4.0",
+    "tailwindcss": "^2.2.19",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.9",
-    "@angular-eslint/builder": "12.3.1",
-    "@angular-eslint/eslint-plugin": "12.3.1",
-    "@angular-eslint/eslint-plugin-template": "12.3.1",
-    "@angular-eslint/schematics": "12.5.0",
-    "@angular-eslint/template-parser": "12.3.1",
-    "@angular/cli": "~12.2.9",
-    "@angular/compiler-cli": "~12.2.9",
+    "@angular-devkit/build-angular": "~13.0.2",
+    "@angular-eslint/builder": "^12.6.1",
+    "@angular-eslint/eslint-plugin": "^12.6.1",
+    "@angular-eslint/eslint-plugin-template": "^12.6.1",
+    "@angular-eslint/schematics": "^12.6.1",
+    "@angular-eslint/template-parser": "^12.6.1",
+    "@angular/cli": "~13.0.2",
+    "@angular/compiler-cli": "~13.0.1",
     "@babel/eslint-plugin": "^7.14.5",
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
+    "@commitlint/cli": "^14.1.0",
+    "@commitlint/config-conventional": "^14.1.0",
+    "@types/jasmine": "~3.10.2",
+    "@types/node": "^12.20.37",
     "@typescript-eslint/eslint-plugin": "4.28.2",
     "@typescript-eslint/parser": "4.28.2",
     "angular-in-memory-web-api": "^0.11.0",
     "eslint": "^7.26.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-ngrx": "^1.0.0",
-    "husky": "^7.0.1",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-ngrx": "^1.46.3",
+    "husky": "^7.0.4",
+    "jasmine-core": "~3.10.1",
+    "karma": "^6.3.8",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2",
-    "standard-version": "^9.3.1",
-    "typescript": "~4.3.5"
+    "lint-staged": "^12.0.2",
+    "prettier": "^2.4.1",
+    "standard-version": "^9.3.2",
+    "typescript": "~4.4.4"
   }
 }
